Use select-outside instead of select-inside in DirDialog
For consistency with FileDialog, where the file that gets selected is
the one under the cursor, DirDialog now also selects the directory
under the cursor.

Previously, the selected directory would have been the one inside which
the user navigated.

This does prevent the user from selecting the root directory. However,
it is not anticipated that the root directory is useful for the sole
current purpose of DirDialog -- selecting the initial directory for a
Selector brought up to ask which file an application gets launched
with.