Place the selection rectangle in the middle of file selectors
BrowseDialog previously set the range of visible files so that the
selected file was always at the bottom of it, unless that was
impossible. Now it sets the range of visible files so that it's in
the middle, unless that's impossible.

It is impossible for the selection rectangle to be in the middle of
the visible range if the selected entry is near the top or bottom
of the entire list of files, or the entire list fits in one page.