Use the last file name selected for a LinkApp as a hint in Selector
This replaces the previous mechanism to preserve the selection, which
was very clunky.

Firstly, it only worked across a single invocation of the Selector and
the subsequent restart of gmenu2x once that application was exited.
Once the user exited the Selector, the last file selected by the user
for that application was forgotten. Now, a 'selectorfile' key in the
application's settings file remembers the name of the last file it was
launched with.

Secondly, because it preserved only the index of a file among the last
selected directory, if the application created a file whose name comes
before the last selected file, the selection would be shifted. This is
mitigated with a file-name-based hint.

Thirdly, it could not recover at all from missing directories, due to
it forcing the selection to an index which could be beyond the file
count. Now, the hint is just applied to the directory that gets used
as a fallback.