Change the subtitle of Selector to be a hint as to what to do
Previously the subtitle would be the description field from the link,
but this is now the text "Select a file to use in this application",
which can be translated.

The motivation for this change is that the description is rarely as
useful as the name of the application. Also, some users were getting
confused as to why a file browser appeared as they launched certain
applications, and tried to provide a program to launch instead of a
data file.