Scrub newlines from file names when displaying them
The POSIX specification for file names allows the use of the newline
character for file names, but Font::write treats it as a control
character. It was possible for a malicious user to place a file whose
name contained a newline character to trick the user into selecting a
file that he or she did not intend to select, or to fill the screen
with garbage text due to overlapping.