Allow the use of the .previews subdirectory in file selectors
For those who can use it, the .previews subdirectory offers an
advantage over using the previews subdirectory: hiding the subdirectory
itself from the file selector, while still showing the images contained
therein.

For those who can't use it, the previews subdirectory can also provide
images.

Windows users may be unable to use .previews due to an oversight in
Windows Explorer which declares an error on file names that start with
a dot, asking for a file name to be entered before the extension.

For completeness, the previews subdirectory has precedence over the
.previews subdirectory. In practice, that will be irrelevant, except if
an application wants to overwrite the image that would have precedence
in future gmenu2x runs (for example, it's about to write a screenshot
of the user's progress in a game, which he or she will see later).